lambdas/ocr_text/requirements.txt
# boto3
Pillow
pytesseract
opencv-python-headless
numpy
